*** empty log message ***

+++ v3.x/kern/msg_clock/clock.c
@@ -117,7 +117,7 @@ static void ClockChrono_Clock(int t)
 	MenageHmsh_tChain11 hourChain;
 	char command;
 	BOOLEAN display;
-	unsigned int msg;
+	unsigned long msg;
 
 	rt_receive(&read, &msg);
 	MenageHmsh_Initialise(&hour);
@@ -164,7 +164,7 @@ static void ClockChrono_Chrono(int t)
 	BOOLEAN display;
 	BOOLEAN hundredthes = FALSE;
 	char command;
-	unsigned int msg;
+	unsigned long msg;
 
 	rt_receive(&read, &msg);
 	command = 'R';
